+++ internal/pkg/group/usecase/get.go
@@ -21,7 +21,7 @@ func (gu *groupUseCase) GetAdministeredGroupByUserId(ctx context.Context, userId
 	return groups, err
 }
 
-func (gu *groupUseCase) CheckIfAdmin(ctx context.Context, userId string, groupId int) (bool, error) {
+func (gu *groupUseCase) CheckIfAdmin(ctx context.Context, userId string, groupId int64) (bool, error) {
 	userIdInt, err := strconv.Atoi(userId)
 	if err != nil {
 		gu.logger.WithError(err).Error("[GetAdministeredGroupByUserId] bad user id")
